pyproject.toml
[tool.poetry]
name = "evals"
version = "0.1.0"
description = ""
authors = ["Belinda Mo <justanexperimentpark@gmail.com>"]
license = "MIT"
readme = "README.md"


# Define where your Python package(s) are
packages = [
    { include = "cover_agent" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
bs4 = "^0.0.2"
fastcore = "^1.5.37"
ghapi = "^1.0.5"
tqdm = "^4.66.4"

# for cover agent
jinja2 = "^3.1.3"
beautifulsoup4 = "^4.12.3"
poetry-dynamic-versioning = "^1.3.0"
litellm = { git = "https://github.com/mrT23/litellm.git" }
openai = "^1.32.0"
tiktoken = "^0.7.0"
boto3 = "^1.34.121"
google-cloud-aiplatform = "^1.54.0"
numpy = "^1.26.0"   # Note: Earlier version have incompatibility issues with google-cloud-aiplatform
dynaconf = "^3.2.4"
wandb = "^0.17.1"
langgraph = "^0.2.12"
langchain-openai = "^0.1.22"
langchain-core = "^0.2.34"
langchain-anthropic = "^0.1.23"

[tool.poetry.group.dev.dependencies]

# for cover agent
pytest = "^8.1.1"
pyinstaller = "^6.6.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"
pytest-timeout = "^2.3.1"
fastapi = "^0.111.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
cover-agent = "cover_agent.main:main"
